requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
mediapipe==0.10.21
opencv-python-headless==4.11.0.86
numpy==2.2.2
Pillow==11.1.0
